[am] Move find_inflight_merges into am.core
Move this functionality into core so it can be reused from the graph
utility.
