+++ git_apple_llvm/am/am_status.py
@@ -3,51 +3,15 @@
 """
 
 from git_apple_llvm.am.am_config import find_am_configs, AMTargetBranchConfig
-from git_apple_llvm.am.core import is_secondary_edge_commit_blocked_by_primary_edge
+from git_apple_llvm.am.core import is_secondary_edge_commit_blocked_by_primary_edge, find_inflight_merges
 from git_apple_llvm.am.oracle import get_ci_status
-from git_apple_llvm.git_tools import git, git_output
+from git_apple_llvm.git_tools import git_output
 import logging
 import click
 from typing import Dict, List, Set, Optional
 
 
-AM_PREFIX = 'refs/am/changes/'
-AM_STATUS_PREFIX = 'refs/am-status/changes/'
-
-
 log = logging.getLogger(__name__)
-
-
-def find_inflight_merges(remote: str = 'origin') -> Dict[str, List[str]]:
-    """
-       This function fetches the refs created by the automerger to find
-       the inflight merges that are currently being processed.
-    """
-    git('fetch', remote,
-        f'{AM_PREFIX}*:{AM_STATUS_PREFIX}*')  # FIXME: handle fetch failures.
-    refs = git_output('for-each-ref', AM_STATUS_PREFIX,
-                      '--format=%(refname)').split('\n')
-
-    inflight_merges: Dict[str, List[str]] = {}
-
-    for ref in refs:
-        if not ref:
-            continue
-        assert ref.startswith(AM_STATUS_PREFIX)
-        merge_name = ref[len(AM_STATUS_PREFIX):]
-        underscore_idx = merge_name.find('_')
-        assert underscore_idx != -1
-        commit_hash = merge_name[:underscore_idx]
-        dest_branch = merge_name[underscore_idx + 1:]
-
-        if dest_branch in inflight_merges:
-            inflight_merges[dest_branch].append(commit_hash)
-        else:
-            inflight_merges[dest_branch] = [commit_hash]
-
-    for (m, k) in inflight_merges.items():
-        log.debug(f'in-flight {m}: {k}')
-    return inflight_merges
 
 
 def compute_inflight_commit_count(commits: List[str], commits_inflight: Set[str]) -> int:
